+++ php/modules/rest-api/postworld-rest-api.php
@@ -187,7 +187,8 @@ class PW_REST_Controller{ // extends WP_REST_Controller
 			if( !$feed )
 				return new WP_Error( 'code', __( 'Feed ID doesn\'t exist', 'postworld' ) );
 			$query = $feed['query'];
-			$result = pw_query( $query, $request['fields'] );
+			$query['fields'] = $request['fields'];
+			return pw_query_posts( $query );
 			return $result->posts;
 		}
 		*/